requirements.txt
langchain
langchain-community
streamlit
langchain-ollama
langchain-qdrant
qdrant-client[fastembed]
pypdf
langchain-text-splitters
python-dotenv
langchain-mistralai